+++ src/main/java/com/plivo/api/models/brand/UseCase.java
@@ -1,0 +1,32 @@
+package com.plivo.api.models.brand;
+
+
+public class UseCase {
+    private String name;
+    private String code;
+    private String details;
+
+    /**
+     * @return getName return the name
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * @return getCode return the code
+     */
+    public String getCode() {
+        return code;
+    }
+
+    /**
+     * @return getDetails return the details
+     */
+    public String getDetails() {
+        return details;
+    }
+
+}
+
+