+++ src/main/java/com/plivo/api/models/campaign/CampaignNumbers.java
@@ -2,6 +2,8 @@ package com.plivo.api.models.campaign;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.plivo.api.models.base.BaseResource;
 
+import java.util.HashMap;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CampaignNumbers extends BaseResource{
    private String campaignID;
@@ -9,6 +11,9 @@ public class CampaignNumbers extends BaseResource{
    private String usecase;
    public PhoneNumbers[] phoneNumbers;
    private String message;
+   public HashMap<String, Integer> phoneNumbersSummary;
+   public Integer numberPoolLimit;
+
 
    public static CampaignNumberLinker link(String campaignID, String url,String method,String[] numbers) {
     return new CampaignNumberLinker(campaignID, numbers,url,method);
@@ -42,6 +47,12 @@ public class CampaignNumbers extends BaseResource{
     }
     public String getMessage(){
         return message;
+    }
+    public HashMap<String, Integer> getPhoneNumbersSummary(){
+        return phoneNumbersSummary;
+    }
+    public Integer getNumberPoolLimit(){
+        return numberPoolLimit;
     }
  @Override
  public String getId() {